src/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
PySocks==1.7.0
pyTelegramBotAPI==3.6.6
python-i18n==0.3.6
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
typing-extensions==3.7.4
urllib3==1.25.3
yarl==1.3.0
